Change directory structure in tests
