+++ tests/Feature/Reader/FilterHandler/AnyHandlerTest.php
@@ -2,14 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Data\Cycle\Tests\Feature\Data\Reader\FilterHandler;
+namespace Yiisoft\Data\Cycle\Tests\Feature\Reader\FilterHandler;
 
 use Yiisoft\Data\Cycle\Exception\NotSupportedFilterException;
-use Yiisoft\Data\Cycle\Tests\Support\NotSupportedFilter;
-use Yiisoft\Data\Reader\Filter\All;
-use Yiisoft\Data\Reader\Filter\Any;
 use Yiisoft\Data\Cycle\Reader\EntityReader;
-use Yiisoft\Data\Cycle\Tests\Feature\Data\BaseData;
+use Yiisoft\Data\Cycle\Tests\Feature\BaseData;
+use Yiisoft\Data\Cycle\Tests\Support\NotSupportedFilter;
+use Yiisoft\Data\Reader\Filter\Any;
 use Yiisoft\Data\Reader\Filter\Equals;
 
 final class AnyHandlerTest extends BaseData
