+++ tests/Unit/Reader/FilterHandler/EqualsHandlerTest.php
@@ -2,23 +2,23 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Data\Cycle\Tests\Unit\Data\Reader\FilterHandler;
+namespace Yiisoft\Data\Cycle\Tests\Unit\Reader\FilterHandler;
 
 use PHPUnit\Framework\TestCase;
 use Yiisoft\Data\Cycle\Exception\UnexpectedFilterException;
-use Yiisoft\Data\Cycle\Reader\FilterHandler\GreaterThanHandler;
+use Yiisoft\Data\Cycle\Reader\FilterHandler\EqualsHandler;
 use Yiisoft\Data\Reader\Filter\Equals;
 use Yiisoft\Data\Reader\Filter\GreaterThan;
 
-final class GreaterThanHandlerTest extends TestCase
+final class EqualsHandlerTest extends TestCase
 {
     public function testUnexpectedFilterException(): void
     {
-        $handler = new GreaterThanHandler();
-        $filter = new Equals('id', 2);
+        $handler = new EqualsHandler();
+        $filter = new GreaterThan('id', 2);
 
         $this->expectException(UnexpectedFilterException::class);
-        $this->expectExceptionMessage(sprintf('Expected "%s", but "%s" given.', GreaterThan::class, Equals::class));
+        $this->expectExceptionMessage(sprintf('Expected "%s", but "%s" given.', Equals::class, GreaterThan::class));
         $handler->getAsWhereArguments($filter, []);
     }
 }
